requirements.txt
pandas>=2.0.0
pyinstaller>=6.0.0
pillow>=10.0.0 